attempt to draw polygons using sprite systems (PolygonSpriteBatch with PolygonRegion). Currently only draws black boxes where shapes should be

+++ core/src/com/mygdx/minigolf/controller/LayerComparator.java
@@ -1,9 +1,7 @@
 package com.mygdx.minigolf.controller;
 
-import com.badlogic.ashley.core.ComponentMapper;
 import com.badlogic.ashley.core.Entity;
 import com.mygdx.minigolf.controller.ComponentMappers.GraphicalMapper;
-import com.mygdx.minigolf.model.components.Graphical;
 
 import java.util.Comparator;
 
@@ -15,6 +13,6 @@ public class LayerComparator implements Comparator<Entity> {
 
     @Override
     public int compare(Entity entityA, Entity entityB) {
-        return Float.compare(GraphicalMapper.get(entityA).getLayer(), GraphicalMapper.get(entityB).getLayer());
+        return Integer.compare(GraphicalMapper.get(entityA).getLayer(), GraphicalMapper.get(entityB).getLayer());
     }
 }
